Add option to dump each step of the simulation to file
This change also changes the setup from a single program that iterates until completion, to one which computes a single step at a time and the iteration is controlled on the Haskell side. This will hurt the CUDA backend since it eagerly copies results back to the host, but PTX should be fine as device->host transfers are done on-demand... I will fix this later.